+++ generators/page/index.js
@@ -19,6 +19,7 @@ module.exports = yeoman.Base.extend({
     this.composeWith('dmninteractives:linters');
     this.composeWith('dmninteractives:common');
     this.composeWith('dmninteractives:nvm');
+    this.composeWith('dmninteractives:gitsecrets');
   },
 
   prompting() {